You can now make konsole a (pseudo)transparent window, just install the two new schemes (Transparent.scheme and GreenTint.scheme) and select them. It works for gradient, tiled and centred backgrounds
TODO:Just have to fix the rest of the "tiled" modes (except tiled itself)
You can define new tints with the transparency setting in the schema file,
just have a look at the examples.

Konsole now generates the background and keeps a cache while it's in the same
desktop. I've had to change it to be a KWMModuleApplication and connect
a few signals for resizes and desktop changes.

I've tried to design it so that it's easy to change if we ever make kbgndwm
put the backgrounds in  shared memory.

svn path=/trunk/kdebase/konsole/; revision=21717