Ok. Konsole now is a part. Try to click in konqy on
kdebase/konsole/src/sample.shell

to see what's happening. Maybe you need to resize konqy, but mc
should appear. (I put mc as a shell, because it is usable with the mouse)
However, the keystrokes do not seem to go through, and if I try to
paste something in, it segfaults. Maybe the "KReadOnlyPart" is at fault?

The integration to konqi should be anyhow done in a special way, so maybe
a konqi hacker can take it from here.

To my best knowledge, I did not break the standalone konsole.

  good night!

svn path=/trunk/kdebase/konsole/; revision=45203