Okay, it does NOT get more bizarre then this... y'know the problem I (and others) was having with konsole not reading in the schema and size anymore? Well, this fixes it for me.
The bizarre part is where the problem appears to be -- it just doesn't
make any sense!

With lots of debug statements, I found out that the 'config' object stopped
reading in *any* values after the 'slotToggleToolBar()' call in
'readProperties'.  That is, the config->readEntry calls would *always*
return the default, ignoring the saved value.

I traced into the slotToggleToolbar function and found that all I had
to do was comment out the 'enableToolbar' function and everything worked
perfectly!

I don't get it at all!  The enableToolbar function has *nothing* to do
with KConfig so how is it changing it all of a sudden :-(

Anyway, I moved the code around in readProperties and it works now.  Very
very very weird.

svn path=/trunk/kdebase/konsole/; revision=46262