Ok. I sort of fixed the resize problem, and the problem with the transparent terminals when launched from kdeinit (the transparent background disappeared when opening the second term).
I think that both of these problems are due to the recent
change from KTMainWindow to KMainWindow.

Resize: before, the resize of the central widget was
reflected in the size of the main window. This is not
the case any more. This was useful because allowed us
to resize the internal widget accurately which is
important if you want to have for example 24 x 80.

I have changed it such that now I am resizing the
external window and try to calculate the size - I am
taking into account the toolbar visibility and position,
menubar visibility etc - but then there are other
problems at startup when the toolbar is not visible etc.

These calculations should be done in the libs not in
the app. Maybe we can revert some of the functionality
in the old ktmainwindow?

Again, the background: krootpixmap installs an event
filter to capture the repaint. To my best knowledge, these
events do not go there if konsole is launched from kdeinit
but they go if it launched from elsewere.

I am 100% sure of this. (I reported this before). And it
is a relatively recent phenomena (several weeks old).

For the time being, I solved the problem by manually
sending a repaint event. It works, but it makes a
flicker when changing terminals.

         Lotzi

svn path=/trunk/kdebase/konsole/; revision=59447