Be gentle when killing processes spawned by a konsole part.
In the past, destroying a konsole part always resulted in a SIGKILL being sent
to the child process (usually a shell). This in turn could have negative effects
such as preventing the shell from writing its history file. This situation was
very painful for users of QuadKonsole.

Now a SIGHUP is sent first. We wait up to 1 second for the process to terminate
before sending a SIGKILL. The method is copied from the non-part konsole.cpp.

And they lived happily ever after.

svn path=/branches/KDE/3.5/kdebase/konsole/; revision=662728