Display a warning message in the terminal in red if starting the session fails because the binary does not exist or if the shell crashes or exits with a non-zero status code.
In these cases, also prevent the tab from being closed automatically.

If the binary for the session cannot be found, display a warning and try
to fall back to $SHELL or as a last attempt, /bin/sh.  This should mean
that users should always be given a semi-functional basic environment
if their profiles get messed up, the shell program gets uninstalled etc.

BUG: 120322

svn path=/trunk/KDE/kdebase/apps/konsole/; revision=787709