Fix consecutive non-word characters being treated as a single word.
TerminalDisplay::charClass() used to return one of three values to indicate whether
the input was a space, "word character" (letter, number or character from the wordCharacters() array)
or something else.  When double clicking a 'word' was taken to be a run of
characters which had the same charClass().  Fix this by returning the input character
in charClass() if it is not a space or word character.

CCBUG:155416

svn path=/trunk/KDE/kdebase/apps/konsole/; revision=791732