Fix calculation of required display size given (lines,columns) in TerminalDisplay widget.  Include width of vertical scrollbar (if visible) and top/left margins.  Konsole will now start with a default width of 80 columns on most systems if no window size has been saved in konsolerc
The initial size of Konsole is specified as 80x40 internally.  Qt constrains
the size of top-level windows on their first automatic resize to a
maximum of 2/3rds of the screen size (see QWidgetPrivate::adjustedSize()).
This means that the display might not get the 80x40 columns that it asks
for, depending on the desktop size.

CCBUG: 160422

svn path=/trunk/KDE/kdebase/apps/konsole/; revision=793964