For remote session, 'Open File Manager' will open the user's home path.
Currently, when using 'Open File Manager' on a remote session,
unexpected results will occur.  For example, for ssh, it will try to
open another Konsole to ssh into the remote server.
Other options to consider in the future, disabling the menu and
transforming URL into what the user might expect.
BUG: 275287