Restore codec to profile setting or previous codec upon reset.
Previously, after typing in 'reset' or using 'Clear Scrollback & Reset',
the current codec is lost.

With this patch, typing in 'reset' will cause the codec to stay what
it was before the reset.  Using the 'Clear Scrollback & Reset' will
cause the codec to be set to the Profile setting.

This may get backported to 4.6.x
BUG: 123866
BUG: 273959
FIXED-IN: 4.7
REVIEW: 101565