Popup menu for drag-n-drop operations.
Restore KDE3 popup menu when any file is dropped into Konsole.

The popup menu allows to Copy, Move or Link the file into the current
working directory or to simply paste its URL as text (the latter being
the current behaviour in KDE4).

For non-local files, the URL is pasted as text.

Note that this adds a konq dependency - KonqOperations:doDrop

Patch by Frank Roscher
FEATURE: 160583
FIXED-IN: 4.8
REVIEW: 4600