Look for LibKonq before using it.
Follow-up to 2dcd5842552684e558716d069b73dd118035b75d.

konsole is not in the same repository as libkonq anymore, so we need
to check if it is installed before using it, and it is also better to
use the whole library path when linking instead of hardcoding "konq".

CCMAIL: konsole-devel@kde.org