Simplify the scrollback setting dialog to avoid potential confusion.
  - the change is not saved into profile and only influences current tab
  - remove the 'Save to current profile' checkbox
  - remove the 'Defaults' button
  - change dialog title from "Scrollback Options" to "Adjust Scrollback"
  - change action from "Scrollback Options..." to "Adjust Scrollback..."

The new dialog makes the "Adjust Scrollback..." action more consistent with most
of other actions in the context menu, since they only influence or work in the
context of current tab.

The UI change is not quite satisfactory. Feel free to refine the UI part
to make it more clear and elegant.

BUG: 227144
FIXED-IN : 4.8
REVIEW: 102688