Change the default shortcut of 'Clear Scrollback and Reset' to Ctrl+Shift+K
The current default shortcut is Ctrl+Shift+X, which might be accidentally
pressed when users intend to press Ctrl+Shift+C for copying or Ctrl+Shift+V for
pasting. Scrollback is important asset, but it is impossible to recover
once gone. So change the default shortcut to decrease the chance of
accidents.

FEATURE: 282593
FIXED-IN: 4.9
REVIEW: 103660