Show the global shortcut for background window in the shortcuts dialog
I'm not sure whether that can be called as improvement. That global
shortcut has always been configurable through systemsettings. Now the
global shortcut is also configurable thourgh shoutcuts dialog, but it
is only available in the background window when it is shown.

I wonder how many users know and really use the --background-mode
option. It feels to me more like a clever trick than a mature feature.