Ensure the 'MenuIndex' entry is never written into profile on disk
The current code expects that entry to be absent or has 0 as its value in the
profile on disk. However, I guess the code is not strict enough to guarantee
that assuption since I have noticed MenuIndex=N several times when checking my
profiles.

More importantly, the way how MenuIndex interacts with favorite status is
not quite staight forward.