Disable the advanced feature of "CommonDirNames" for "%d" by default
The main consideration is it is more often confusing than helpful for
average users, especially when this feature was hidden, undocumented and
enabled by default.

The "CommonDirNames" list in the code is now empty by default , instead
of a list of hardcoded common names.

To enable and customize this advanced and potentally confusing feature,
edit $KDEHOME/usr/share/config/konsolerc and add the following lines:

    [ProcessInfo]
    CommonDirNames=name1,name2,name3...

For konsolepart users, edit the config file of the hosting application.

I'll add related information for this hidden & advanced feature into
handbook later.

BUG:190281
FIXED-IN:4.9.0

REVIEW: 103949
CCMAIL: robertknight@gmail.com