Add workaround for the problem that Session::processId() might return 0
That was not an actual problem before KDE 4.8, because the old code
guaranteed when Session::processId() was called, the session had
already started running so it wouldn't return 0.

Since KDE 4.8, that guarantee was not there any more since a hard-coded
200ms delay is introduced to avoid the session starting "too quickly".
In short, this is another side effect of the fix for infameous BKO #176902.

See https://bugzilla.novell.com/show_bug.cgi?id=745954 for one concrete
issue caused by this underlying problem