Remove the long broken --force-transparency option
That option is broken for a long time (at least since KDE 4.6.2), yet
there is not a single bug report against it. That is a strong evince
that option is seldom used.

The ideal situation is whenever compositing is available, Konsole
supports translucent background out of box without user guidance.