Make sure the translucent background is available whenever possible
Now the Qt::WA_TranslucentBackground attribute is always set for the
mainwindow of konsole(regardless of whether compositing is available
when the window is created), just like in konsolepart.

I'm not 100% sure this approach is problem-free, but it works for me so
far under various environments and conditions. Do not hesihate to revert
it if you notice it causes big trouble.