Add cmdline options for the visibility of menubar and tabbar
Since the menubar and tarbar settings are now global instead of per
profile , those cmdline options are useful to allow users to override
the defautl behavior.

The --hide-menubar and --hide-tabbar options can be used together
to simulate the miminal interface of xterm. I'm not sure whether
--show-menubar and --show-tabbar will be widely used , but add them at
the moment for the sake of completeness.