When saving profile, always use its untranslated name as filename
An internal property is introduced to record the untranslated name.
When saving profile into disk, always refer to that untranslated name
to get the filename.

This extra consideration is intended to deal with the situation when the
profile's contents are i18nized. The typical example is the "Shell" profile
shipped by default.

BUG: 250506
FIXED-IN: 4.9.0