Expand %# %w in the "Copy Input To" dialog session list.
Duplicate the code in SessionController.cpp to expand %# %w in the
copy input dialog.  This is not the ideal solution but it not trivial
to fix w/ all the tab/window titles issues in Konsole.

BUG: 175755
FIXED-IN: 4.10
(cherry picked from commit f43dd28d66b2a78157f1b5d14727f8542be9126a)