Specify $HOME as working directory through the .desktop file
This is just a workaround for the problem of opening "~/Documents" on startup.
Although it is simple (just one line) and easy to understand, it is still a
not-so-pleasant hack which doesn't need to be added at the first place.

Note, the per-profile working directory still takes precedence if it is set

CCBUG: 302903