fix broken quoting in konsoleprofile
Add missing quotes around '$1' in konsoleprofile so that we do not
perform word splitting twice (which is very unlikely to be what the user
wants, and is difficult or impossible to prevent). This fixes a
regression introduced in 4e1b96e4.

BUG: 307351
FIXED-IN: 4.9.2
(cherry picked from commit 7d05ceedf80a140b465b1b35851f7179ddd89e2e)