Remove the hack of resetting working directory to $HOME
That hack is added due to the the "single process" model used by konsole.
However, That hack seems to make the thing more confusing as reflected in bug
302903

So here is a known issue: if you set konsole as the terminal emulator
used in pcmanfm, pressing F4 twice in different folder will open two
konsole window in the same folder.