Revert "Fix comment typo"
This reverts commit 726c52c4cf7683385205a6ffac4c9fae529dd706.

This comes as one part of reverting all the work related with the
refactoring of the ssh command parser. That big refactoring commit
5a5fd236 causes some hard to debug crashes/problems. For the long term
benefit, reverting now and early is a good choice. I will try to redo the
refactoring in a better way for KDE SC 4.11, but for the upcoming 4.10
stability more important.