Revert "Refacotr and move the logic of parsing ssh command into a separate function"
This reverts commit 5a5fd236963af12ed319dd75cb848e12f971b47b.

This comes as one part of reverting all the work related with the
refactoring of the ssh command parser. That big refactoring commit
5a5fd236 causes some hard to debug crashes/problems. For the long term
benefit, reverting now and early is a good choice. I will try to redo the
refactoring in a better way for KDE SC 4.11, but for the upcoming 4.10
stability more important.