Add Dialog warning when huge amounts of text is about to be pasted
Currently there is no way to stop an accidently paste into the terminal.
This patch adds a dialog when 8000+ characters are pasted asking if
the user really wants to do this.  At most the user will see this once
as there is a "Don't show again" box.

Previously commit trigger a sysadmin issue.

FEATURE: 162625
REVIEW: 107570
FIXED-IN: 4.11
GUI: