Use QRect instead of manually handling terminal margins
Instead of keeping track of constants and +1 -1, use a QRect to handle
the terminal margins.  There is no GUI for this currently and the
margins are hard-coded to what they are now.  Code to center the
terminal is also include (w/o GUI).

This patch comes from review 109891 w/ a few changes from me.
Thanks to Mariusz Glebocki mglb@arccos-1.net