Fix crash by not assuming standard shortcuts are present.
Neither of the two standard shortcuts used here are *guaranteed* to be
present (kdelibs filters both with some KAuthorized magic, and
help_contents also depends on the documentation subsystem), so avoid
crashing just because they're not around.

First noticed on my Gentoo-provided Konsole which is compiled without
the handbook USE flag.