ProcessInfo: isolate all the platform-specific *ProcessInfo implementations
compile all the platform-specific *ProcessInfo implementations only
in the OS they belong to, basically extending this conditional
compilation also to the Linux and Solaris implementations.

This allows to use proper OS-specific API/etc in each, removing quirks
(like the one for the Solaris psinfo struct) to emulate such APIs on
other OSes.