Fix double click can only select text within visible region
Double-clicking on text will highlight text not on the visible screen.
This is the 2nd part of this patch.  The first part for triple clicking
highlighting the entire line is already in 2.12.

Thanks to Yichao Yu yyc1992@gmail.com for this big patch.

REVIEW: 111441
BUG: 319260
FIXED-IN: 2.13
(cherry picked from commit 0b69bdcd40d93d9b87b729a978974b1cc2a77b35)