Restore --background-mode
For now restore this option that was removed for 2.11.  This option
is somewhat a hack and abnormal feature.  However, it seems a number
of people use it and I can see how it would be useful.  So instead of
spending a lot of time right now trying to correct the issues, just
restore the previous code.
Ideally for KDE 5, this portion of the code can be better handled.

35bb9cf9e79465d252580615f50b4ce8d34e3d34

BUG: 320783
FIXED-IN: 2.13#