Allow users to specify css file for tab bar style
Currently, the default stylesheet for the tab bar just sets the min
and max width.  Now the user can load a .css file.  It is up to the
user to generate the css file to their liking.  Handbook documentation
and examples will follow.

This can be used to set minimum width of the tabs, distinguish active
tab, etc...
BUG: 330252
BUG: 330558
BUG: 330556
FEATURE:
REVIEW: 115554
FIXED-IN: 2.13