Fix rows calculation for new profile settings.
New feature in 2.13 allows columns/rows in profiles; however, the
rows would be incorrect depending on if tabbar was visible.  Ideally,
this should not be needed but I would guess this won't get fixed
correctly until frameworks timeframe.
This calculation can still be wrong if ShowNavigationAsNeeded and
perhaps other times depending on height of tabbar.

FIXED: 2.13