Issue a warning when not matching a font exactly
When setting a font for the terminal, if it is not found properly by the
system, Qt attempts to use another font which is not an exact match. It
happens that the other font is not fixed-pitch, so konsole ends up using
an unsupported variable-pitch font for the terminal.

Place a warning when an exact match for the desired font has not been
found.