Fix storing of profile when no profile loaded.
If for some reason no profile is loaded, i. e. only the built-in
fallback default values are used, we need to create a new profile to
ensure that changes to the profile are actually persisted.

This has worked more or less by chance for most people so far, because
the default is to load the Shell.profile that is shipped with Konsole.
But if for example the konsolerc config file has an empty value for
DefaultProfile it will load the builtin fallback profile, which can't be
saved.

BUG: 342188
REVIEW: 124142