Allow user to select where the scrollback/history files are written.
Currently, when unlimited scrollback is used, the files are stored in
/tmp.  As noted in the bug report, this can cause issues on some
distros.  The default is still /tmp but the user can selected to use
a location in their $HOME folder (QStandardPaths::CacheLocation).

This is somewhat a WIP.  Allowing for the changes to take effect
immediately appears to be a gigantic mess to get to work.

CCBUG: 173283