Allow certain variable-width fonts to be used
For certain fonts (with ligatures) QFontInfo::fixedPitch doesn't return
the correct result.  Previously, these fonts were not allowed to be
used.  Variable-width fonts are still not selectable in the font dialog.

Thanks to Stefan Seemayer stefan seemayer de for research and patch.

See reviewboard for a more detailed description

REVIEW: 126921