Fix crash when KDBusService exits
KDBusService is designed in a way that is fundamentally not compatible
with Qt, where it exit()s before the QApplication is deleted, which
leads to a lot of weird crashes in various parts of Qt.

To work around this we install an atexit() handler to delete the
QApplication before we shut down.

REVIEW: 127626