Show note about and allow removing sendText/runCommand from D-Bus
There are some security concerns about allowing sendText and runCommand
to be on the public D-Bus.  Now, once upon a Konsole startup, a
notification will popup noting this concern when either of these two
methods are called.  There is a cmake definition that will remove
these methods from the public D-Bus.

Discussion on the KDE security mailing list.