add a debian compat option
debian policy for reasons beyond my apprehension wants a -T option to
set the window title (like in xterm).
to make debian's (and derivate's) life easier let's support a no-op
-T option but hide it. this option will not appear in --help output
but be accepted by the parser all the same.

this enables Debian and friends to drop a long standing useless patch
and at the same time prevents us from introducing a conflicting -T in
the future.

REVIEW: 128299