Preserve current line when resetting
Reset was always called with clearScreen = true, so no point in having
that variable. We can probably also remove Screen::clear() and
Screen::home(), they're not used anymore, and weren't really useful
before this either.

We still preserve the scrollback history when resetting, which is not
what other terminal emulators do (http://stackoverflow.com/a/5367075),
but we do have a separate action for it, so I think it is a nice extra
feature to preserve it.

REVIEW: 128389