Improve filter updating
This should update filters less often, and also hopefully fix the issue
where underlined links get stuck while scrolling.

The idea is to not to do it on all mouse updates, only when needed on
mouse updates, and also not display the filters if they are out of date.
The only thing that should happen on screen changes now is setting the
_filterUpdateRequired variable to true, which leads to the filters not
being displayed until re-processed, which happens the next time the
mouse moves, unless I fucked something up.

REVIEW: 128393