Allow non monospaced fonts to be selected
Due to various issues outside the scope of Konsole, some monospaced
fonts are not allowed to be selected.  The default is still to only
allow system found monospaced fonts but users can bypass to allow any
font to be picked. There is a warning about issues when doing this.

CCBUG: 349350