Fix scrolling down if user presses something which is not just a modifier
If the user presses e. g. arrow up, we should actually scroll down. What
I intended to fix was only the case where the user only pressed a
modifier.

And hardcoding a list of keys and behaviors is bad and bug prone, so I
think this is the best solution.

REVIEW: 128827