Add --nofork as compatibility alias for --separate
In previous incarnations of kuniqueapplication it used to inject a common
command option --nofork which is meant to bypass single-instance behavior.
Given that konsole can and is being invoked from scripts they may well want
to ensure that the fork they created is the actual instance of konsole.
i.e. to monitor return values and life time

Presently, since the options are divergent between konsole4 and konsole5,
scripts are either incompatible with older konsoles or with newer konsoles.
To make life easier for everyone add a compat alias --nofork, which
behaves exactly like separate.

(this unbreaks steam, which is a notable recent offender of falling into
 this particular trap)

REVIEW: 129647
CHANGELOG: Added compatibility command option `--nofork` for single-instance behavior