Restore zoom fonts after changing profile with ESC sequences
Summary:
When a profile property is changed via the command line or an ESC sequence
a new temp profile is created and applied, this resets the zoom values in
the views connected to the session.
Save zoom font sizes and restore them after the new profile is applied.

BUG: 386643
FIXED-IN: 18.04

Reviewers: hindenburg

Reviewed By: hindenburg

Subscribers: ngraham, #konsole

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D10650