Konsole: Add read-only mode
Summary:
This patch adds a read-only option for TerminalDisplays. When active,
all keyboard events are eaten. Mouse input is not affected and works like before.
The setting is not persisted and only lasts for the duration of the session.

Screenshots:
{F5730786}
{F5730788}
{F5730789}

VDG input is highly appreciated. Also, I'm not sure if I can just change the rc-files without bumping the version?

FEATURE: 126930

Test Plan:
- Shortcuts still work
- Paste and drop actions are disabled when readonly
- Switching between read-only and normal tabs works as expected
- Mouse input works like before

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: hindenburg, ngraham, #konsole

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D10862