Apply the key bindings changes to the current profile
Summary:
For the Profile::KeyBindings property, if it's set in the _tempProfile
then the user opened the edit key bindings dialog and clicked OK, and
could have add/removed a key bindings rule. Allow applying the settings
so that the current profile KeyBindings property is updated without
having to restart konsole or open a new tab.

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: #konsole

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D12255