Paste dropped files and folders as text by default
Summary:
When a user drags a file or folder onto the Konsole window, by far the
most common and likely thing they want to happen is for the path to
that item appear in the window. This is what virtually all other terminal
programs do by default. Konsole does not, and I suspect 99% of its
users do not know that this setting can be turned off, as I did not know
until I started digging through the code itself.

This patch makes the paste-as-text behavior happen by default for new
profiles. Super advanced users can re-enable the drag menu if they like it.
For most Konsole users, I think not having it presents vastly better
drag-and-drop usability and increases the speed of hybrid mouse+gui
operations. As such, I believe this is an important change for {T6831}.

I know changing defaults like this can be scary and controversial, but I
really do believe that the "paste as text" behavior is what 99% of drags
to konsole are intended for. It's not at all like Dolphin where a user may
reasonably want to copy or link instead of move.

FEATURE: 393523
FIXED-IN 18.08.0

Test Plan: Make a new user, open Konsole, drag a file or folder to the window; it's immediately pasted as text

Reviewers: #konsole, hindenburg, #vdg

Reviewed By: #konsole, hindenburg

Subscribers: abetts, hein, #plasma

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D12529