Fallback to fixed width fonts
Summary:
Before this commit when Konsole was trying to render a character that
didn't exist in the font selected by user, it'd fall back to default font provided
by font config without any hints that it needs monospace fonts which usually
would be a sans font.

Now Konsole hints system that it needs a monospace font.

BUG: 393620

Reviewers: hindenburg

Reviewed By: hindenburg

Subscribers: #konsole, hindenburg, mglb, ahmadsamir

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D12614