Fix cursor advancement in last column
Summary:
Until now, the cursor was moved from last column to the next column
outside the screen. Since screen array is linear, it partially worked
like it was on the beginning of a next line (as long as it was not the
last line).

Test Plan:
* Run `cat > /dev/null` (shells do their own line wrapping, so will
  not work)
* Move cursor to terminal's last visible line using enter
* Move cursor to terminal's last visible column by typing something
* Type one more character

Expected result: new line should appear with the cursor on the beginning
Actual result: cursor disappears, no new line

* Turn on blinking cursor
* Run `cat > /dev/null`
* Move cursor to terminal's last visible column by typing something
* Type one more character

Expected result: the cursor should move to new line's first column
and blink
Actual result: the cursor "moves" but does not blink

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: konsole-devel, hindenburg, #konsole

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D12554