Reduce the boilerplate needed for commandLineArgs
Summary:
Use the new initialization syntax plus a initialization vector
to remove the need to call parser->addOption(QCommandLineOption
on every single command line option.

The code now store things in a vector, that is in fact slower
but this should be negligible, and the code is easier to extend
and to read.

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: hindenburg, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D12966