Remove StackedViewContainer class
Summary:
StackedViewContainer is a plain view container with no navigiations.
It can be replaced using TabbedViewContainer and hiding the tabbar.
This will minimize duplicate code and allow easier control of code
in Konsole and KonsolePart.

Test Plan:
Run konsole and konsole part test, and then
use the shortcuts to create tab.

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: pedroarthurp, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D12964