Ignore control characters in the text part of Xpe "ESC]" sequences
Summary:
Ignore control characters in the text part of Xpe (OSC, Operating System
Control) of the form "ESC]Pn;Pt\a" escape sequences; this matches what
XTerm docs say.

If userTitle is an empty string, set the window title to a blank space
when showWindowTitleOnTitleBar is enabled; basically if the user sets
the title to an empty string, leave it as-is.

BUG: 315581
FIXED-IN: 18.08

Test Plan:
- unset PROMPT_COMMAND, in case the shell is configured to set the
   window title via that variable
- Make sure "Show window title on the title bar" is enabled
- Execute:
$ /usr/bin/echo -e 'j\e]2;\n\aj'
- The window title is changed to the tab title text; and the output
    looks like this:
j
j

- Apply the diff, build, and try again, the window title should be
   set to an empty string (really a blank space), and the output
   should look like:
jj

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: konsole-devel, #konsole

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D13078